+++ packages/terra-core-docs/src/terra-dev-site/test/overlay/LoadingOverlay/AnimatedLoadingOverlay.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import OverlayContainer from '../../../../OverlayContainer';
-import LoadingOverlay from '../../../../LoadingOverlay';
+import OverlayContainer from 'terra-overlay/lib/OverlayContainer';
+import LoadingOverlay from 'terra-overlay/lib/LoadingOverlay';
 import styles from '../Overlay/OverlayTestCommon.module.scss';
 
 const cx = classNames.bind(styles);