+++ packages/terra-core-docs/src/terra-dev-site/test/overlay/Overlay/FullscreenScrollableOverlay.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Overlay from '../../../../Overlay';
+import Overlay from 'terra-overlay/lib/Overlay';
 
 class OverlayExample extends React.Component {
   constructor() {