+++ packages/terra-core-docs/CHANGELOG.md
@@ -22,6 +22,7 @@
   * Migrate docs from terra-i18n.
   * Migrate docs from terra-icon.
   * Migrate docs from terra-image.
+  * Migrate docs from terra-overlay.
 
 ## 1.12.0 - (November 2, 2021)